+++ src/helpers/interfaces/firestorm/textures.ts
@@ -11,3 +11,11 @@ export interface Texture {
   contributions?: Contributions;
 }
 export interface Textures extends Array<Texture> {}
+
+export interface MCMETA {
+  animation: {
+    interpolate?: true;
+    frametime?: number;
+    frames?: Array<number | { index: number; time: number }>;
+  };
+}